+++ src/example/java/io/github/computerdaddyguy/jfiletreeprettyprinter/example/ChildLimitDynamic.java
@@ -1,16 +1,15 @@
 package io.github.computerdaddyguy.jfiletreeprettyprinter.example;
 
-import io.github.computerdaddyguy.jfiletreeprettyprinter.ChildLimitBuilder;
+import io.github.computerdaddyguy.jfiletreeprettyprinter.ChildLimits;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.FileTreePrettyPrinter;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.PathMatchers;
 
 public class ChildLimitDynamic {
 
 	public static void main(String[] args) {
-		var isNodeModuleMatcher = PathMatchers.hasName("node_modules");
-		var childLimit = ChildLimitBuilder.builder()
-			.defaultLimit(ChildLimitBuilder.UNLIMITED)
-			.limit(isNodeModuleMatcher, 0)
+		var childLimit = ChildLimits.builder()
+			.setDefault(ChildLimits.UNLIMITED) // Unlimited children by default
+			.add(PathMatchers.hasName("node_modules"), 0) // Do NOT print any children in "node_modules" folder
 			.build();
 		var prettyPrinter = FileTreePrettyPrinter.builder()
 			.customizeOptions(options -> options.withChildLimit(childLimit))