+++ src/example/java/io/github/computerdaddyguy/jfiletreeprettyprinter/example/ProjectStructure.java
@@ -1,9 +1,10 @@
 package io.github.computerdaddyguy.jfiletreeprettyprinter.example;
 
-import io.github.computerdaddyguy.jfiletreeprettyprinter.ChildLimitBuilder;
+import io.github.computerdaddyguy.jfiletreeprettyprinter.ChildLimits;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.FileTreePrettyPrinter;
+import io.github.computerdaddyguy.jfiletreeprettyprinter.LineExtensions;
 import io.github.computerdaddyguy.jfiletreeprettyprinter.PathMatchers;
-import io.github.computerdaddyguy.jfiletreeprettyprinter.PrettyPrintOptions.Sorts;
+import io.github.computerdaddyguy.jfiletreeprettyprinter.PathSorts;
 import java.nio.file.Path;
 import java.util.Comparator;
 import java.util.function.Function;
@@ -47,46 +48,38 @@ public class ProjectStructure {
 		var fileFilter = PathMatchers.allOf(
 
 			// Hide files with names starting with "."
-			PathMatchers.not(PathMatchers.hasNameStartingWith(".")),
-
-			// Inside "jfiletreeprettyprinter" folder, keep only "FileTreePrettyPrinter.java"
-			// Files in other folders are not restricted by this rule.
-			PathMatchers.ifMatchesThenElse(
-				/*  if  */ PathMatchers.hasDirectParentMatching(PathMatchers.hasName("jfiletreeprettyprinter")),
-				/* then */ PathMatchers.hasName("FileTreePrettyPrinter.java"),
-				/* else */ path -> true
-			)
+			PathMatchers.not(PathMatchers.hasNameStartingWith("."))
 		);
 
 		/*
 		 * Limit the number of displayed children by directory: some content is not relevant and clutters the final result!
 		 */
-		var childLimitFunction = ChildLimitBuilder.builder()
+		var childLimitFunction = ChildLimits.builder()
 			// Hide all files under renderer and scanner packages
-			.limit(PathMatchers.hasAbsolutePathMatchingGlob("**/io/github/computerdaddyguy/jfiletreeprettyprinter/renderer"), 0)
-			.limit(PathMatchers.hasAbsolutePathMatchingGlob("**/io/github/computerdaddyguy/jfiletreeprettyprinter/scanner"), 0)
+			.add(PathMatchers.hasAbsolutePathMatchingGlob("**/io/github/computerdaddyguy/jfiletreeprettyprinter/renderer"), 0)
+			.add(PathMatchers.hasAbsolutePathMatchingGlob("**/io/github/computerdaddyguy/jfiletreeprettyprinter/scanner"), 0)
+			.add(PathMatchers.hasAbsolutePathMatchingGlob("**/io/github/computerdaddyguy/jfiletreeprettyprinter"), 3)
 			.build();
 
 		/*
 		 * Add some comments on a few files and directories
 		 */
-		Function<Path, String> lineExtension = path -> {
-			if (PathMatchers.hasName("project-structure.png").matches(path)) {
-				return "\t// This image";
-			} else if (PathMatchers.hasName("FileTreePrettyPrinter.java").matches(path)) {
-				return "\t// Main entry point";
-			} else if (PathMatchers.hasName("README.md").matches(path)) {
-				return "\t\t// You're reading at this!";
-			} else if (PathMatchers.hasRelativePathMatchingGlob(projectFolder, "src/main/java").matches(path)) {
-				return ""; // Empty string: force line break in compact directory chain
-			}
-			return null;
-		};
+		Function<Path, String> lineExtension = LineExtensions.builder()
+			.add(PathMatchers.hasName("project-structure.png"), "\t// This image")
+			.add(PathMatchers.hasName("FileTreePrettyPrinter.java"), "\t// Main entry point")
+			.add(PathMatchers.hasName("README.md"), "\t\t// You're reading at this!")
+			.addLineBreak(PathMatchers.hasRelativePathMatchingGlob(projectFolder, "src/main/java"))
+			.build();
 
 		/*
-		 * Sort all paths by directory first (then alphabetically by default)
+		 * Sort all paths by directory first (with highest precedence),
+		 * then "FileTreePrettyPrinter.java" has precedence "-100".
+		 * All other files have default precedence "0", and are then sorted alphabetically by default.
 		 */
-		Comparator<Path> pathComparator = Sorts.DIRECTORY_FIRST;
+		Comparator<Path> pathComparator = PathSorts.builder()
+			.addFirst(PathMatchers.isDirectory())
+			.add(PathMatchers.hasName("FileTreePrettyPrinter.java"), -100) // Default precedence is "0"
+			.build();
 
 		/*
 		 * Build the final FileTreePrettyPrinter
@@ -115,25 +108,27 @@ public class ProjectStructure {
 		        Expected result
 		 ================================
 		 
-			📂 JFileTreePrettyPrinter/
-			├─ 📂 assets/
-			│  └─ 🖼️ project-structure.png	// This image
-			├─ 📂 src/main/java/
-			│  └─ 📂 io/github/computerdaddyguy/jfiletreeprettyprinter/
-			│     ├─ 📂 renderer/
-			│     │  └─ ... (5 files and 2 directories skipped)
-			│     ├─ 📂 scanner/
-			│     │  └─ ... (4 files skipped)
-			│     └─ ☕ FileTreePrettyPrinter.java	// Main entry point
-			├─ 🗺️ CHANGELOG.md
-			├─ 📖 CONTRIBUTING.md
-			├─ 📄 LICENSE
-			├─ 📖 README.md		// You're reading at this!
-			├─ 🗺️ ROADMAP.md
-			├─ 🛡️ SECURITY.md
-			├─ 🏗️ pom.xml
-			├─ 📖 release_process.md
-			└─ 📜 runMutationTests.sh
+		📂 JFileTreePrettyPrinter/
+		├─ 📂 assets/
+		│  └─ 🖼️ project-structure.png	// This image
+		├─ 📂 src/main/java/
+		│  └─ 📂 io/github/computerdaddyguy/jfiletreeprettyprinter/
+		│     ├─ 📂 renderer/
+		│     │  └─ ... (5 files and 2 directories skipped)
+		│     ├─ 📂 scanner/
+		│     │  └─ ... (4 files skipped)
+		│     ├─ ☕ FileTreePrettyPrinter.java	// Main entry point
+		│     └─ ... (10 files skipped)
+		├─ 🗺️ CHANGELOG.md
+		├─ 📖 CONTRIBUTING.md
+		├─ 📄 LICENSE
+		├─ 📖 README.md		// You're reading at this!
+		├─ 🗺️ ROADMAP.md
+		├─ 🛡️ SECURITY.md
+		├─ 🏗️ pom.xml
+		├─ 📖 release_process.md
+		└─ 📜 runMutationTests.sh
+		
 		 */
 	}